Fixed several performance issues
The game lagged pretty badly already when reaching level 2 or 3, so some changes were necessary. First off the spawning cylinders and corresponding animations weren't deleted after appearing, same for the corresponding materials and the healthbar material. Performances were still not good in later levels, turns out that creating a new GlowLayer for each healthbar is not a good idea. Deleted also the sun's GlowLayer which affected performances even in the first level. Also added fps to screen. At last, I centered the camera on the player instead of the center of the planet